system/tools/requirements.txt
requests==2.19.0
pyyaml==5.3.1
flask==0.12
